Stream the heavy asset pages and collapse their query waterfalls
The location and ship pages each ran roughly ten Supabase round trips
one after another, most of which had no reason to wait on each other:
the character and corp probes for the same item were serialized, the
owner list was fetched last despite depending on nothing, and the type
name was looked up twice over.

Fires the independent work concurrently instead, cutting both pages from
about ten sequential round trips to three or four, and folds two lookups
into one where the same SDE read already answered both questions (a
container's name and whether it is a ship come off the same row).

Splits each page at its real seam. What runs before the first byte is
only what the response depends on -- the auth and share checks, the root
items, which location this is, and whether the id names a ship, since
that decides whether the page is a redirect at all. The expensive work
now hangs off a Suspense boundary: the per-container subtree counts, the
whole-hangar location sweep a system page does, and the ship's module
table and fit wheel. Their fallbacks reuse phase 1's skeletons, so a
streamed-in table and a freshly-navigated one look alike.

Also overlaps /fitting's owner resolution with its hull lookup, which
keys off the fits rather than the owners.

Phase 2 of docs/page-load-performance.md. No behavior or data changes.
