+++ src/app/asset/[locationId]/page.tsx
@@ -1,66 +1,156 @@
+import type { SupabaseClient } from '@supabase/supabase-js'
 import { notFound, redirect } from 'next/navigation'
+import { Suspense } from 'react'
 
 import { getSdeSystem } from '@/sdeSystems'
-import { getSdeTypes } from '@/sdeTypes'
+import { getSdeType } from '@/sdeTypes'
 import { createClient } from '@/utils/supabase/server'
 import { createServiceClient } from '@/utils/supabase/service'
 import { AssetPath, fetchAssetPath, regionSystemCrumbs, type Crumb } from '../../assetPath'
-import { typeFacts } from '../../assetTypeFacts'
 import { fetchOwners } from '../../owners'
-import { resolveLocations, type LocationRef } from '../../resolveLocations'
+import type { Owners } from '../../ownerFilter'
+import { SkeletonTable } from '../../skeleton'
+import { fetchStationNames, fetchStationSystems } from '../../stationNames'
+import { fetchSystemNames } from '../../systemNames'
+import { TypeIcon } from '../../typeIcon'
 import { resolveShareParams } from '../access'
 import { saveAssetShare, revokeAssetShare } from '../shareActions'
 import { fetchShareDialogData } from '../shareData'
 import { ShareDialog } from '../shareDialog'
-import { fetchStationNames, fetchStationSystems } from '../../stationNames'
-import { fetchSystemNames } from '../../systemNames'
-import { TypeIcon } from '../../typeIcon'
-import { fetchTypeNames } from '../../typeNames'
-import type { Owners } from '../../ownerFilter'
 import styles from '../assets.module.css'
-import { type Location } from '../assetsTable'
 import { AppraisalPanel } from './appraisalPanel'
-import { LocationAssets, type ItemRow } from './locationAssets'
-import { SystemLocations } from './systemLocations'
-
-// A row from the *_asset_location_summary() RPCs, normalized to whoever owns the
-// stacks (a character registration uuid or an EVE corporation id).
-type SummaryRow = {
-  location_id: number | string
-  location_type: string | null
-  owner_id: string
-  stacks: number | string
-}
-
-// invGroups.categoryID for the Ship category in CCP's SDE.
-const SHIP_CATEGORY_ID = 6
-
-// Bigint ids arrive from PostgREST as strings; keep them as strings and only
-// convert at the API/system-lookup boundary (mirrors the assets index page).
-type CharacterAsset = {
-  item_id: number | string
-  registration_id: string
-  type_id: number | string
-  location_id: number | string | null
-  location_flag: string | null
-  location_type: string | null
-  quantity: number | string | null
-  is_singleton: boolean | null
-  is_blueprint_copy: boolean | null
-  name: string | null
-}
-
-// Corp assets carry no player-assigned name column; owner is the corporation.
-type CorpAsset = Omit<CharacterAsset, 'registration_id' | 'name'> & { corporation_id: number | string }
-
-// Either source's row, normalized to whoever owns it: a character
-// (registration uuid) or a corporation (EVE corporation id).
-type Asset = Omit<CharacterAsset, 'registration_id'> & { owner_id: string }
+import { SHIP_CATEGORY_ID, fetchRootItems, type CharacterAsset, type CorpAsset } from './locationData'
+import { LocationItems } from './locationItems'
+import { SystemLocationsSection } from './systemLocationsSection'
 
 type Structure = {
   structure_id: number | string
   name: string | null
   system_id: number | string | null
+}
+
+// The location itself, when the id turns out to be one of our own items rather
+// than a place — a ship or container the user drilled into.
+type Self = Pick<CharacterAsset, 'item_id' | 'type_id' | 'location_id' | 'name'>
+
+// Whether this id is one of the caller's own items, and whose. Both tables are
+// probed at once and the character row wins: an item lives in exactly one of
+// them, so the extra lookup is a cheap indexed miss rather than a second
+// round trip waiting on the first to come back empty.
+const fetchSelf = async (
+  supabase: SupabaseClient,
+  locationId: string,
+  characterScope: string[] | null,
+  corpScope: number[] | null
+): Promise<{ self: Self | null; isOwnCharacterItem: boolean }> => {
+  let selfQuery = supabase
+    .from('character_asset')
+    .select('item_id, registration_id, type_id, location_id, name')
+    .eq('item_id', locationId)
+  if (characterScope) selfQuery = selfQuery.in('registration_id', characterScope)
+  let corpSelfQuery = supabase
+    .from('corp_asset')
+    .select('item_id, corporation_id, type_id, location_id')
+    .eq('item_id', locationId)
+  if (corpScope) corpSelfQuery = corpSelfQuery.in('corporation_id', corpScope)
+
+  const [{ data: characterSelf }, { data: corpSelf }] = await Promise.all([
+    selfQuery.maybeSingle<Pick<CharacterAsset, 'item_id' | 'registration_id' | 'type_id' | 'location_id' | 'name'>>(),
+    corpSelfQuery.maybeSingle<Pick<CorpAsset, 'item_id' | 'corporation_id' | 'type_id' | 'location_id'>>(),
+  ])
+
+  return {
+    self: characterSelf ?? (corpSelf ? { ...corpSelf, name: null } : null),
+    isOwnCharacterItem: characterSelf != null,
+  }
+}
+
+// Everything the heading needs when the id is a place rather than an item:
+// which structure/station/system it is, and the system it sits in. Resolved the
+// same way the index page does — own-corp + foreign player structures, NPC
+// stations, and systems all come from the universe_name DB cache.
+//
+// The four lookups are independent of each other, so they go out together; only
+// the system *name* has to wait, since which system to name depends on them.
+const fetchPlace = async (
+  supabase: SupabaseClient,
+  locationId: string,
+  corpScope: number[] | null,
+  // The location_type the root items report, when there are any. A location
+  // holding items already says what kind of place it is.
+  locationType: string | null
+) => {
+  const numericId = Number(locationId)
+
+  let corpStructureQuery = supabase
+    .from('corp_structure')
+    .select('structure_id, name, system_id')
+    .eq('structure_id', numericId)
+  if (corpScope) corpStructureQuery = corpStructureQuery.in('corporation_id', corpScope)
+
+  const [{ data: corpStructures }, { data: playerStructures }, stationNames, stationSystems, sdeSystemProbe] =
+    await Promise.all([
+      corpStructureQuery,
+      supabase.from('universe_structure').select('structure_id, name, system_id').eq('structure_id', numericId),
+      fetchStationNames([numericId]),
+      fetchStationSystems([numericId]),
+      // A system id resolves in the SDE; a station/structure id doesn't. Only
+      // needed when the type is unknown (no root items) — the case that needs
+      // disambiguating: a system opened directly with nothing floating in its
+      // space. Fired alongside the structure lookups rather than after them (it
+      // is a process-cached SDE read, so an unused probe costs nothing) and
+      // discarded below if this turns out to be a structure.
+      locationType == null && Number.isFinite(numericId) ? getSdeSystem(numericId) : null,
+    ])
+
+  const structure = (corpStructures?.[0] ?? playerStructures?.[0] ?? null) as Structure | null
+  const sdeSystem = structure ? null : sdeSystemProbe
+  const isSolarSystem = !structure && (locationType === 'solar_system' || sdeSystem != null)
+
+  const systemId = isSolarSystem
+    ? numericId
+    : structure?.system_id != null
+      ? Number(structure.system_id)
+      : stationSystems[numericId]
+  const systemNames = await fetchSystemNames(
+    [systemId, isSolarSystem ? numericId : undefined].filter((n): n is number => n != null)
+  )
+
+  const heading =
+    (structure
+      ? (structure.name ?? `Structure #${locationId}`)
+      : isSolarSystem
+        ? (systemNames[numericId] ?? sdeSystem?.name ?? `System #${locationId}`)
+        : stationNames[numericId]) ?? `Location #${locationId}`
+
+  return {
+    heading,
+    isSolarSystem,
+    systemId,
+    systemName: systemId != null ? (systemNames[systemId] ?? `#${systemId}`) : undefined,
+  }
+}
+
+// The share-token path has no session to read owner names with, so they come
+// from the resolved scope through the service client instead.
+const fetchScopeOwners = async (
+  supabase: SupabaseClient,
+  scope: NonNullable<Awaited<ReturnType<typeof resolveShareParams>>>
+): Promise<Owners> => {
+  const characters = [...scope.characterNames]
+    .map(([id, name]) => ({ id, name }))
+    .sort((a, b) => a.name.localeCompare(b.name))
+  let corpNames: Record<number, string> = {}
+  if (scope.corporationIds.length > 0) {
+    const { data } = await supabase.from('universe_name').select('id, name').in('id', scope.corporationIds)
+    corpNames = Object.fromEntries(
+      ((data ?? []) as Array<{ id: number | string; name: string }>).map((r) => [Number(r.id), r.name])
+    )
+  }
+  const corporations = scope.corporationIds
+    .map((id) => ({ id: String(id), name: corpNames[id] ?? `Corporation #${id}` }))
+    .sort((a, b) => a.name.localeCompare(b.name))
+  return { characters, corporations }
 }
 
 // Directory browsing for a location's contents. Ships are their own thing —
@@ -70,6 +160,13 @@ type Structure = {
 // client, so every query is explicitly filtered to the sharing user's
 // characters/corps — a shared station view must never reveal other users'
 // items parked in the same station.
+//
+// The page is deliberately split in two. What runs before the first byte is
+// only what the answer depends on: the share/auth check, the root items, who
+// this location is, and — because it decides whether this page is a redirect at
+// all — whether the id names a ship. Everything expensive (the per-container
+// subtree counts, and on a system page the whole-hangar location sweep) hangs
+// off a Suspense boundary below, so the heading and breadcrumb don't wait on it.
 const AssetLocationPage = async ({
   params,
   searchParams,
@@ -96,193 +193,68 @@ const AssetLocationPage = async ({
   const characterScope = scope?.registrationIds ?? null
   const corpScope = scope ? (scope.corporationIds.length > 0 ? scope.corporationIds : [-1]) : null
 
-  // Root-level assets at this location: ships, containers and loose stacks that
-  // sit directly in the station/structure/system (not nested in another item),
-  // from both the character and corp hangars. Only this level is fetched — the
-  // whole asset table no longer pages into Node.
-  let characterQuery = supabase
-    .from('character_asset')
-    .select(
-      'item_id, registration_id, type_id, location_id, location_flag, location_type, quantity, is_singleton, is_blueprint_copy, name'
-    )
-    .eq('location_id', locationId)
-  if (characterScope) characterQuery = characterQuery.in('registration_id', characterScope)
-  let corpQuery = supabase
-    .from('corp_asset')
-    .select(
-      'item_id, corporation_id, type_id, location_id, location_flag, location_type, quantity, is_singleton, is_blueprint_copy'
-    )
-    .eq('location_id', locationId)
-  if (corpScope) corpQuery = corpQuery.in('corporation_id', corpScope)
-  // A character's currently-piloted ship reports its location as wherever it's
-  // docked — physically indistinguishable from any other ship parked there —
-  // so it otherwise shows up as an ordinary item in that station's listing.
-  // Still shown (it belongs here from the DB's point of view), but flagged so
-  // the UI can tell it apart from a ship that's merely parked.
-  let currentShipQuery = supabase.from('character_ship').select('ship_item_id')
-  if (characterScope) currentShipQuery = currentShipQuery.in('registration_id', characterScope)
-  const [{ data: characterChildren }, { data: corpChildren }, { data: currentShips }] = await Promise.all([
-    characterQuery,
-    corpQuery,
-    currentShipQuery,
+  // Nothing here depends on anything else here, so it all goes out at once.
+  // These used to run one after another, each waiting on a round trip that had
+  // no reason to wait.
+  const [{ rootItems, currentShipItemIds }, { self, isOwnCharacterItem }, owners] = await Promise.all([
+    fetchRootItems(supabase, locationId, characterScope, corpScope),
+    fetchSelf(supabase, locationId, characterScope, corpScope),
+    // Owner names for the filter dropdown: the caller's own owners, or — in the
+    // token path — the sharer's characters/corps, resolved through the service
+    // client (the anonymous viewer has no session to read them with).
+    scope ? fetchScopeOwners(supabase, scope) : fetchOwners(),
   ])
-  const currentShipItemIds = new Set(
-    ((currentShips ?? []) as { ship_item_id: number | string }[]).map((s) => String(s.ship_item_id))
-  )
-  const rootItems: Asset[] = [
-    ...((characterChildren ?? []) as CharacterAsset[]).map(({ registration_id, ...a }) => ({
-      ...a,
-      owner_id: registration_id,
-    })),
-    ...((corpChildren ?? []) as CorpAsset[]).map(({ corporation_id, ...a }) => ({
-      ...a,
-      owner_id: String(corporation_id),
-      name: null,
-    })),
-  ]
 
-  // Total items held inside each ship/container at this location, counted across
-  // the whole subtree by the *_asset_location_contents() functions in Postgres.
-  // An item lives in exactly one of the two tables, so the maps can't collide.
-  // Skipped in the token path: those functions walk every asset unscoped when
-  // run as service_role, and a shared hangar doesn't offer drill-down anyway.
-  const contentsByItem = new Map<string, number>()
-  if (!scope) {
-    const [{ data: characterContents }, { data: corpContents }] = await Promise.all([
-      supabase.rpc('character_asset_location_contents', { parent: locationId }),
-      supabase.rpc('corp_asset_location_contents', { parent: locationId }),
-    ])
-    for (const r of [...(characterContents ?? []), ...(corpContents ?? [])] as {
-      item_id: number | string
-      contents: number | string
-    }[]) {
-      contentsByItem.set(String(r.item_id), Number(r.contents))
-    }
-  }
-
-  // The id is either a place (station / structure / solar system) or one of our
-  // own items — a ship or container the user drilled into, character- or
-  // corp-owned. Resolve the heading and the "back" target accordingly.
-  let selfQuery = supabase
-    .from('character_asset')
-    .select('item_id, registration_id, type_id, location_id, name')
-    .eq('item_id', locationId)
-  if (characterScope) selfQuery = selfQuery.in('registration_id', characterScope)
-  const { data: characterSelf } =
-    await selfQuery.maybeSingle<
-      Pick<CharacterAsset, 'item_id' | 'registration_id' | 'type_id' | 'location_id' | 'name'>
-    >()
-  let corpSelfQuery = supabase
-    .from('corp_asset')
-    .select('item_id, corporation_id, type_id, location_id')
-    .eq('item_id', locationId)
-  if (corpScope) corpSelfQuery = corpSelfQuery.in('corporation_id', corpScope)
-  const { data: corpSelf } = characterSelf
-    ? { data: null }
-    : await corpSelfQuery.maybeSingle<Pick<CorpAsset, 'item_id' | 'corporation_id' | 'type_id' | 'location_id'>>()
-  const self = characterSelf ?? (corpSelf ? { ...corpSelf, name: null } : null)
-
-  // Share dialog data for an owned character item; fetchShareDialogData
-  // returns null unless the caller actually owns it (visibility alone can now
-  // mean "shared with me").
-  const shareData = !scope && characterSelf ? await fetchShareDialogData(supabase, locationId) : null
-
-  // One bulk SDE lookup for every type in play (this location's root items plus
-  // the location itself, if it's an item). It feeds three things: the set of
-  // Ship-category type ids, so the per-row ship test stays a sync Set.has; each
-  // row's volume/group/category columns; and which image variation its icon has.
-  const itemTypes = await getSdeTypes([
-    ...rootItems.map((a) => Number(a.type_id)),
-    ...(self ? [Number(self.type_id)] : []),
-  ])
-  const shipTypeIds = new Set(
-    Object.values(itemTypes)
-      .filter((t) => t.categoryID === SHIP_CATEGORY_ID)
-      .map((t) => t.typeID)
-  )
-  const isShip = (typeId: number | string) => shipTypeIds.has(Number(typeId))
+  // The self type answers two questions at once — what to call this container,
+  // and whether it's a ship — off one process-cached SDE read.
+  const selfType = self ? await getSdeType(Number(self.type_id)) : null
 
   // A ship is its own page, not a directory level (a share link stays valid
   // through the redirect — a signed link covers the whole shared subtree, a
   // legacy token is bound to this same id).
-  if (self && isShip(self.type_id)) {
+  if (self && selfType?.categoryID === SHIP_CATEGORY_ID) {
     redirect(`/ship/${locationId}${share ? `?share=${share}` : token ? `?token=${token}` : ''}`)
   }
 
-  let heading: string
+  // Share dialog data for an owned character item; fetchShareDialogData returns
+  // null unless the caller actually owns it (visibility alone can now mean
+  // "shared with me"). Started rather than awaited: whether the item is ours was
+  // settled above, so this overlaps the heading work below instead of queueing
+  // behind it.
+  const shareDataPromise = !scope && isOwnCharacterItem ? fetchShareDialogData(supabase, locationId) : null
+
   // Where this location lives: the full container chain up to its root place,
   // rendered as the breadcrumb above the heading. A bare place (station /
   // structure / system) has no ancestry, so its path is just the Assets root.
-  // Skipped in the token path: the RPC would walk unscoped as service_role,
-  // and a shared hangar shouldn't reveal where it lives anyway.
+  // The ancestry RPC is skipped in the token path: it would walk unscoped as
+  // service_role, and a shared hangar shouldn't reveal where it lives anyway.
+  let heading: string
   let crumbs: Crumb[] = [{ href: '/asset', label: 'Assets' }]
   let systemName: string | undefined
   // True when the id is a solar system: its heading is the system name, and it
-  // lists the stations/structures in the system where we hold assets (see
-  // systemLocations below) rather than only the items floating in its space.
+  // lists the stations/structures in the system where we hold assets rather
+  // than only the items floating in its space.
   let isSolarSystem = false
-  let systemLocations: Location[] = []
+  let systemId: number | null = null
 
   if (self) {
-    // A container: title it by its custom name (if any) plus type.
-    const typeNames = await fetchTypeNames([Number(self.type_id)])
-    const typeName = typeNames[Number(self.type_id)] ?? `#${self.type_id}`
+    // A container: title it by its custom name (if any) plus type. The type
+    // name is already in hand from the redirect check above, so naming it costs
+    // no round trip of its own.
+    const typeName = selfType?.name ?? `#${self.type_id}`
     heading = self.name && self.name !== typeName ? `${self.name} (${typeName})` : typeName
-    if (!scope) {
-      crumbs = await fetchAssetPath(locationId, supabase)
-    }
+    if (!scope) crumbs = await fetchAssetPath(locationId, supabase)
   } else {
-    // Resolve the location's own name/system the same way the index page does:
-    // own-corp + foreign player structures, NPC stations, and systems all come
-    // from the universe_name DB cache.
-    const numericId = Number(locationId)
-    let corpStructureQuery = supabase
-      .from('corp_structure')
-      .select('structure_id, name, system_id')
-      .eq('structure_id', numericId)
-    if (corpScope) corpStructureQuery = corpStructureQuery.in('corporation_id', corpScope)
-    const { data: corpStructures } = await corpStructureQuery
-    const { data: playerStructures } = await supabase
-      .from('universe_structure')
-      .select('structure_id, name, system_id')
-      .eq('structure_id', numericId)
-    const structure = (corpStructures?.[0] ?? playerStructures?.[0] ?? null) as Structure | null
-
-    const locationType = rootItems[0]?.location_type ?? null
-    // A system id resolves in the SDE; a station/structure id doesn't. Only
-    // probed when the type is unknown (no root items) — the case that needs
-    // disambiguating: a system opened directly with nothing floating in its
-    // space. A location with items already reports its type.
-    const sdeSystem =
-      locationType == null && !structure && Number.isFinite(numericId) ? await getSdeSystem(numericId) : null
-    isSolarSystem = !structure && (locationType === 'solar_system' || sdeSystem != null)
-
-    const [stationNames, stationSystems] = await Promise.all([
-      fetchStationNames([numericId]),
-      fetchStationSystems([numericId]),
-    ])
-
-    const systemId = isSolarSystem
-      ? numericId
-      : structure?.system_id != null
-        ? Number(structure.system_id)
-        : stationSystems[numericId]
-    const systemNames = await fetchSystemNames(
-      [systemId, isSolarSystem ? numericId : undefined].filter((n): n is number => n != null)
-    )
-
-    heading =
-      (structure
-        ? (structure.name ?? `Structure #${locationId}`)
-        : isSolarSystem
-          ? (systemNames[numericId] ?? sdeSystem?.name ?? `System #${locationId}`)
-          : stationNames[numericId]) ?? `Location #${locationId}`
-    systemName = systemId != null ? (systemNames[systemId] ?? `#${systemId}`) : undefined
+    const place = await fetchPlace(supabase, locationId, corpScope, rootItems[0]?.location_type ?? null)
+    heading = place.heading
+    isSolarSystem = place.isSolarSystem
+    systemId = place.systemId ?? null
+    systemName = place.systemName
 
     // Breadcrumb: region › system for this place (the place itself is the
     // heading). For a bare solar system the system is the heading, so only the
     // region shows. Falls back to the Assets root when it can't be placed.
-    const prefix = await regionSystemCrumbs(systemId, systemName, isSolarSystem)
+    const prefix = await regionSystemCrumbs(place.systemId, place.systemName, place.isSolarSystem)
     if (prefix.length > 0) {
       crumbs = prefix
       // The system now lives in the breadcrumb (region › system), so drop the
@@ -290,103 +262,22 @@ const AssetLocationPage = async ({
       // fallback, where the breadcrumb can't show the system.
       systemName = undefined
     }
-
-    // For a solar system, the items parked in its stations/structures live under
-    // those locations (location_id = station/structure), not the system id, so
-    // the root-item query only surfaced things floating in space. List every
-    // station/structure in this system where we hold assets as a navigable
-    // directory, built from the same RLS-scoped location-summary RPCs the index
-    // page uses. Authenticated path only — those RPCs walk unscoped as
-    // service_role, and a share token is always bound to a single item.
-    if (isSolarSystem && !scope) {
-      const [{ data: characterSummary }, { data: corpSummary }] = await Promise.all([
-        supabase.rpc('character_asset_location_summary'),
-        supabase.rpc('corp_asset_location_summary'),
-      ])
-      const summary: SummaryRow[] = [
-        ...((characterSummary ?? []) as Array<Omit<SummaryRow, 'owner_id'> & { registration_id: string }>).map(
-          ({ registration_id, ...r }) => ({ ...r, owner_id: registration_id })
-        ),
-        ...((corpSummary ?? []) as Array<Omit<SummaryRow, 'owner_id'> & { corporation_id: number | string }>).map(
-          ({ corporation_id, ...r }) => ({ ...r, owner_id: String(corporation_id) })
-        ),
-      ]
-      // Tally stacks per location/owner, dropping the system's own space bucket
-      // (its items are already shown inline below, and it would self-link).
-      const byLocation = new Map<string, { root: LocationRef; counts: Map<string, number> }>()
-      for (const row of summary) {
-        const id = String(row.location_id)
-        if (id === locationId) continue
-        const entry = byLocation.get(id) ?? { root: { id, type: row.location_type }, counts: new Map() }
-        entry.counts.set(row.owner_id, (entry.counts.get(row.owner_id) ?? 0) + Number(row.stacks))
-        byLocation.set(id, entry)
-      }
-      const entries = [...byLocation.values()]
-      const { nameFor, systemIdFor } = await resolveLocations(
-        entries.map(({ root }) => root),
-        supabase
-      )
-      systemLocations = entries
-        .filter(({ root }) => systemIdFor(root) === numericId)
-        .map(({ root, counts }) => ({
-          id: root.id,
-          name: nameFor(root),
-          system: null,
-          counts: Object.fromEntries(counts),
-        }))
-    }
   }
 
-  // Owner names for the filter dropdown: the caller's own owners, or — in the
-  // token path — the sharer's characters/corps, resolved through the service
-  // client (the anonymous viewer has no session to read them with).
-  let owners: Owners
-  if (scope) {
-    const characters = [...scope.characterNames]
-      .map(([id, name]) => ({ id, name }))
-      .sort((a, b) => a.name.localeCompare(b.name))
-    let corpNames: Record<number, string> = {}
-    if (scope.corporationIds.length > 0) {
-      const { data } = await supabase.from('universe_name').select('id, name').in('id', scope.corporationIds)
-      corpNames = Object.fromEntries(
-        ((data ?? []) as Array<{ id: number | string; name: string }>).map((r) => [Number(r.id), r.name])
-      )
-    }
-    const corporations = scope.corporationIds
-      .map((id) => ({ id: String(id), name: corpNames[id] ?? `Corporation #${id}` }))
-      .sort((a, b) => a.name.localeCompare(b.name))
-    owners = { characters, corporations }
-  } else {
-    owners = await fetchOwners()
-  }
+  const shareData = shareDataPromise ? await shareDataPromise : null
 
-  // Resolve type names without blocking render: fire the lookup and let each
-  // TypeName stream in (Suspense), falling back to #id. A big hangar can hold
-  // hundreds of distinct types, and awaiting them all here times the page out.
-  const typeNamesPromise = fetchTypeNames(rootItems.map((a) => Number(a.type_id)))
-
-  // Containers/ships (those holding items) first, then a stable id order.
-  // Anonymous hangar viewers get no drill-down links: nested levels would
-  // need their own tokens.
-  const rows: ItemRow[] = rootItems
-    .map((a) => {
-      const contents = contentsByItem.get(String(a.item_id)) ?? 0
-      const type = itemTypes[Number(a.type_id)]
-      return {
-        itemId: String(a.item_id),
-        ownerId: a.owner_id,
-        typeId: Number(a.type_id),
-        name: a.name,
-        quantity: a.quantity,
-        isSingleton: a.is_singleton,
-        flag: a.location_flag,
-        contents,
-        isCurrentShip: currentShipItemIds.has(String(a.item_id)),
-        href: scope ? null : isShip(a.type_id) ? `/ship/${a.item_id}` : contents > 0 ? `/asset/${a.item_id}` : null,
-        ...typeFacts(type, a.is_blueprint_copy),
-      }
-    })
-    .sort((a, b) => b.contents - a.contents || a.typeId - b.typeId)
+  const items = (
+    <Suspense fallback={<ItemsFallback />}>
+      <LocationItems
+        supabase={supabase}
+        locationId={locationId}
+        rootItems={rootItems}
+        currentShipItemIds={currentShipItemIds}
+        owners={owners}
+        linkable={!scope}
+      />
+    </Suspense>
+  )
 
   return (
     <>
@@ -424,20 +315,36 @@ const AssetLocationPage = async ({
 
       {isSolarSystem ? (
         <>
-          <SystemLocations locations={systemLocations} owners={owners} />
+          {!scope && systemId != null ? (
+            <Suspense fallback={<SystemLocationsFallback />}>
+              <SystemLocationsSection supabase={supabase} locationId={locationId} systemId={systemId} owners={owners} />
+            </Suspense>
+          ) : null}
           {/* Items floating directly in the system's space (not docked in any of
               its stations) — usually none, so only shown when present. */}
           {rootItems.length > 0 ? (
             <>
               <h2 className="serif">In space</h2>
-              <LocationAssets rows={rows} owners={owners} typeNamesPromise={typeNamesPromise} canAppraise={!scope} />
+              {items}
             </>
           ) : null}
         </>
       ) : (
-        <LocationAssets rows={rows} owners={owners} typeNamesPromise={typeNamesPromise} canAppraise={!scope} />
+        items
       )}
     </>
   )
 }
 export default AssetLocationPage
+
+// Same shape the route's loading.tsx uses, so a streamed-in table and a
+// freshly-navigated one look like the same page at the same stage.
+const ItemsFallback = () => (
+  <SkeletonTable
+    columns={['Quantity', 'Item', 'Volume (m³)', 'Group', 'Category', 'Owner', 'Hangar', 'Contents']}
+    numeric={['Quantity', 'Volume (m³)', 'Contents']}
+    rows={8}
+  />
+)
+
+const SystemLocationsFallback = () => <SkeletonTable columns={['Location', 'Items']} numeric={['Items']} rows={5} />
