+++ src/app/ship/[itemId]/page.tsx
@@ -1,15 +1,15 @@
+import type { SupabaseClient } from '@supabase/supabase-js'
 import { notFound, redirect } from 'next/navigation'
-import { ascend, sortWith } from 'ramda'
+import { Suspense } from 'react'
 
-import { getSdeType, getSdeTypes, type SdeType } from '@/sdeTypes'
+import { getSdeType, getSdeTypes } from '@/sdeTypes'
 import { createClient } from '@/utils/supabase/server'
 import { createServiceClient } from '@/utils/supabase/service'
 import { AssetPath, fetchAssetPath } from '../../assetPath'
 import { typeFacts } from '../../assetTypeFacts'
-import { fetchOwners } from '../../owners'
 import type { Owners } from '../../ownerFilter'
+import { SkeletonTable } from '../../skeleton'
 import { fetchTypeNames } from '../../typeNames'
-import { flagSortKey } from '../../fitting/fit'
 import { AppraisalPanel } from '../../asset/[locationId]/appraisalPanel'
 import { LocationAssets, type ItemRow } from '../../asset/[locationId]/locationAssets'
 import { resolveShareParams } from '../../asset/access'
@@ -17,83 +17,21 @@ import { saveAssetShare, revokeAssetShare } from '../../asset/shareActions'
 import { fetchShareDialogData } from '../../asset/shareData'
 import { ShareDialog } from '../../asset/shareDialog'
 import { toEsiFit } from './esfFit'
+import { FitPlaceholder } from './fitPlaceholder'
 import { characterPortrait, corporationLogo, ShipHeading, type ShipOwner } from './shipHeading'
+import { ShipContents } from './shipContents'
 import { ShipFitViewDynamic } from './shipFitViewDynamic'
-
-// invGroups.categoryID for the Ship category in CCP's SDE.
-const SHIP_CATEGORY_ID = 6
+import { SHIP_CATEGORY_ID, fittingOrder, hullRow, type ChildRow } from './shipRows'
 
 type ShipRow = {
   item_id: number | string
   type_id: number | string
-  location_id: number | string | null
-  location_type: string | null
+  location_id?: number | string | null
+  location_type?: string | null
   name: string | null
   registration_id?: string
   corporation_id?: number | string
 }
-
-type ChildRow = {
-  item_id: number | string
-  type_id: number | string
-  location_flag: string | null
-  quantity: number | string | null
-  is_singleton: boolean | null
-  is_blueprint_copy?: boolean | null
-  name?: string | null
-  registration_id?: string
-  corporation_id?: number | string
-}
-
-// Whether a single type is in the Ship category — one SDE lookup, for the
-// self-type gate that decides ship page vs asset browser.
-const isShipType = async (typeId: number | string): Promise<boolean> =>
-  (await getSdeType(Number(typeId)))?.categoryID === SHIP_CATEGORY_ID
-
-// Fitting-window order for the module table: slot family first (high → mid →
-// low → rigs → bays, via the same flagSortKey the fitting pages use), the flag
-// string within a family (HiSlot0 before HiSlot1), then type id. This is the
-// server order the unsorted table shows, so the listing reads top-to-bottom
-// the way the ship does.
-const fittingOrder = sortWith<ItemRow>([
-  ascend((r) => flagSortKey(r.flag ?? '')),
-  ascend((r) => r.flag ?? ''),
-  ascend((r) => r.typeId),
-])
-
-// The hull as a row of its own, heading the module table.
-//
-// It isn't one of the ship's children — it's their container — so nothing in
-// the child query produces it, and the table used to list a ship's contents
-// without listing the ship. That reads as an omission now that the table is
-// what gets selected and appraised: "everything shown in this list" has to
-// include the thing the page is about. Assembled, hence a singleton with no
-// stack size, and it links nowhere: this is already its page.
-const hullRow = (
-  self: Pick<ShipRow, 'item_id' | 'type_id' | 'name'>,
-  ownerId: string,
-  type: SdeType | undefined,
-  contents: number
-): ItemRow => ({
-  itemId: String(self.item_id),
-  ownerId,
-  typeId: Number(self.type_id),
-  name: self.name ?? null,
-  quantity: null,
-  isSingleton: true,
-  flag: null,
-  contents,
-  isCurrentShip: false,
-  href: null,
-  ...typeFacts(type),
-})
-
-// How many items the hull holds, all the way down: its own children plus
-// whatever each of them contains. The per-child counts come from the same
-// *_location_contents() walk the drill-in links use, so the hull's Contents
-// cell agrees with the rows under it.
-const nestedCount = (children: ChildRow[], contentsByItem: Map<string, number>): number =>
-  children.length + children.reduce((total, c) => total + (contentsByItem.get(String(c.item_id)) ?? 0), 0)
 
 // A ship's own page: the eveship.fit wheel + stats built from its fitted
 // modules, a full module/cargo table in fitting order (the same sortable
@@ -121,143 +59,46 @@ const ShipPage = async ({
     redirect('/')
   }
 
-  // The ship row, from whichever hangar owns it.
-  const { data: characterSelf } = await supabase
-    .from('character_asset')
-    .select('item_id, registration_id, type_id, location_id, location_type, name')
-    .eq('item_id', itemId)
-    .maybeSingle<ShipRow>()
-  const { data: corpSelf } = characterSelf
-    ? { data: null }
-    : await supabase
-        .from('corp_asset')
-        .select('item_id, corporation_id, type_id, location_id, location_type')
-        .eq('item_id', itemId)
-        .maybeSingle<ShipRow>()
-  const self = characterSelf ?? corpSelf
-  if (!self) notFound()
-  // Non-ships (containers, loose stacks) live on the asset browser instead.
-  if (!(await isShipType(self.type_id))) redirect(`/asset/${itemId}`)
-
-  const [{ data: characterChildren }, { data: corpChildren }] = await Promise.all([
+  // The ship row, from whichever hangar owns it. Both hangars are probed at
+  // once and the character row wins: an item lives in exactly one of them, so
+  // the corp lookup is a cheap indexed miss rather than a second round trip
+  // waiting on the first to come back empty.
+  const [{ data: characterSelf }, { data: corpSelf }] = await Promise.all([
     supabase
       .from('character_asset')
-      .select('item_id, registration_id, type_id, location_flag, quantity, is_singleton, is_blueprint_copy, name')
-      .eq('location_id', itemId),
+      .select('item_id, registration_id, type_id, location_id, location_type, name')
+      .eq('item_id', itemId)
+      .maybeSingle<ShipRow>(),
     supabase
       .from('corp_asset')
-      .select('item_id, corporation_id, type_id, location_flag, quantity, is_singleton, is_blueprint_copy')
-      .eq('location_id', itemId),
+      .select('item_id, corporation_id, type_id, location_id, location_type')
+      .eq('item_id', itemId)
+      .maybeSingle<ShipRow>(),
   ])
-  const children = [...((characterChildren ?? []) as ChildRow[]), ...((corpChildren ?? []) as ChildRow[])]
+  const self = characterSelf ?? corpSelf
+  if (!self) notFound()
 
-  // Nested-contents counts drive the drill-in links on container/ship rows.
-  const [{ data: characterContents }, { data: corpContents }] = await Promise.all([
-    supabase.rpc('character_asset_location_contents', { parent: itemId }),
-    supabase.rpc('corp_asset_location_contents', { parent: itemId }),
-  ])
-  const contentsByItem = new Map<string, number>(
-    (
-      [...(characterContents ?? []), ...(corpContents ?? [])] as {
-        item_id: number | string
-        contents: number | string
-      }[]
-    ).map((r) => [String(r.item_id), Number(r.contents)])
-  )
+  // One process-cached SDE read answers both questions this page opens with:
+  // whether the id is a ship at all, and what to call it.
+  const selfType = await getSdeType(Number(self.type_id))
+  // Non-ships (containers, loose stacks) live on the asset browser instead.
+  if (selfType?.categoryID !== SHIP_CATEGORY_ID) redirect(`/asset/${itemId}`)
 
-  // Owner: the holding character's name and portrait, or the corporation's
-  // cached name and logo. `character_id` here is the EVE numeric id (what the
-  // image server serves portraits for), not the registration uuid its
-  // sibling asset columns misname.
-  let owner: ShipOwner
-  if (characterSelf?.registration_id) {
-    const { data: registration } = await supabase
-      .from('registration')
-      .select('name, character_id')
-      .eq('id', characterSelf.registration_id)
-      .maybeSingle<{ name: string; character_id: number | string | null }>()
-    // A shared ship's owner is outside the caller's registration view (RLS);
-    // their public name resolves through the world-readable directory.
-    const { data: directory } = registration
-      ? { data: null }
-      : await supabase
-          .from('character_directory')
-          .select('name, character_id')
-          .eq('registration_id', characterSelf.registration_id)
-          .maybeSingle<{ name: string | null; character_id: number | string | null }>()
-    const eveCharacterId = registration?.character_id ?? directory?.character_id ?? null
-    owner = {
-      name: registration?.name ?? directory?.name ?? 'Unknown character',
-      portrait: eveCharacterId == null ? null : characterPortrait(eveCharacterId),
-    }
-  } else {
-    const corporationId = Number(corpSelf?.corporation_id)
-    const { data: corpName } = await supabase
-      .from('universe_name')
-      .select('name')
-      .eq('id', corporationId)
-      .maybeSingle<{ name: string }>()
-    owner = { name: corpName?.name ?? `Corporation #${corporationId}`, portrait: corporationLogo(corporationId) }
-  }
-
-  // Where the ship lives: the full container chain up to its station /
-  // structure / system, rendered as the breadcrumb above the heading.
-  const crumbs = await fetchAssetPath(itemId, supabase)
-
-  const typeNames = await fetchTypeNames([Number(self.type_id)])
-  const typeName = typeNames[Number(self.type_id)] ?? `#${self.type_id}`
+  const typeName = selfType?.name ?? `#${self.type_id}`
   const heading = self.name && self.name !== typeName ? `${self.name} (${typeName})` : typeName
 
-  // One bulk SDE lookup → set of Ship-category type ids among the cargo, so the
-  // per-row drill-in link test stays a sync Set.has. The hull's own type rides
-  // along for its row's volume/group/category columns.
-  const childTypes = await getSdeTypes([Number(self.type_id), ...children.map((c) => Number(c.type_id))])
-  const childShipTypeIds = new Set(
-    Object.values(childTypes)
-      .filter((t) => t.categoryID === SHIP_CATEGORY_ID)
-      .map((t) => t.typeID)
-  )
-
-  const typeNamesPromise = fetchTypeNames([Number(self.type_id), ...children.map((c) => Number(c.type_id))])
-  const rows: ItemRow[] = fittingOrder(
-    children.map((c) => {
-      const contents = contentsByItem.get(String(c.item_id)) ?? 0
-      return {
-        itemId: String(c.item_id),
-        ownerId: c.registration_id ?? String(c.corporation_id),
-        typeId: Number(c.type_id),
-        name: c.name ?? null,
-        quantity: c.quantity,
-        isSingleton: c.is_singleton,
-        flag: c.location_flag,
-        contents,
-        isCurrentShip: false,
-        href: childShipTypeIds.has(Number(c.type_id))
-          ? `/ship/${c.item_id}`
-          : contents > 0
-            ? `/asset/${c.item_id}`
-            : null,
-        ...typeFacts(childTypes[Number(c.type_id)], c.is_blueprint_copy),
-      }
-    })
-  )
-
-  // The hull heads the table; the fit viewer below is fed the children alone,
-  // since a ship isn't a module fitted to itself.
-  const tableRows: ItemRow[] = [
-    hullRow(
-      self,
-      characterSelf?.registration_id ?? String(corpSelf?.corporation_id),
-      childTypes[Number(self.type_id)],
-      nestedCount(children, contentsByItem)
-    ),
-    ...rows,
-  ]
-
-  // The share dialog appears only for a character item the caller actually
-  // owns — with phase 2's widening policy, RLS visibility alone can also mean
-  // "shared with me", which must not offer the dialog.
-  const [shareData, owners] = await Promise.all([fetchShareDialogData(supabase, itemId), fetchOwners()])
+  // Who this ship is and where it sits — all independent of each other, and of
+  // the contents below, so they go out together.
+  const [owner, crumbs, shareData] = await Promise.all([
+    fetchShipOwner(supabase, characterSelf, corpSelf),
+    // The full container chain up to its station / structure / system,
+    // rendered as the breadcrumb above the heading.
+    fetchAssetPath(itemId, supabase),
+    // The share dialog appears only for a character item the caller actually
+    // owns — with phase 2's widening policy, RLS visibility alone can also mean
+    // "shared with me", which must not offer the dialog.
+    fetchShareDialogData(supabase, itemId),
+  ])
 
   return (
     <>
@@ -283,10 +124,71 @@ const ShipPage = async ({
           </>
         }
       />
-      <ShipFitViewDynamic esiFit={toEsiFit(Number(self.type_id), self.name ?? null, rows)} />
-      <LocationAssets rows={tableRows} owners={owners} typeNamesPromise={typeNamesPromise} canAppraise />
+      <Suspense fallback={<ContentsFallback />}>
+        <ShipContents
+          supabase={supabase}
+          itemId={itemId}
+          self={self}
+          ownerId={characterSelf?.registration_id ?? String(corpSelf?.corporation_id)}
+        />
+      </Suspense>
     </>
   )
+}
+
+// While the child query and its subtree walk are in flight: the fit viewer's
+// own reserved footprint (it has a second, client-side wait after this) above a
+// stand-in for the module table.
+const ContentsFallback = () => (
+  <>
+    <FitPlaceholder />
+    <SkeletonTable
+      columns={['Quantity', 'Item', 'Volume (m³)', 'Group', 'Category', 'Owner', 'Hangar', 'Contents']}
+      numeric={['Quantity', 'Volume (m³)', 'Contents']}
+      rows={8}
+    />
+  </>
+)
+
+// Owner: the holding character's name and portrait, or the corporation's
+// cached name and logo. `character_id` here is the EVE numeric id (what the
+// image server serves portraits for), not the registration uuid its sibling
+// asset columns misname.
+const fetchShipOwner = async (
+  supabase: SupabaseClient,
+  characterSelf: ShipRow | null,
+  corpSelf: ShipRow | null
+): Promise<ShipOwner> => {
+  if (!characterSelf?.registration_id) {
+    const corporationId = Number(corpSelf?.corporation_id)
+    const { data: corpName } = await supabase
+      .from('universe_name')
+      .select('name')
+      .eq('id', corporationId)
+      .maybeSingle<{ name: string }>()
+    return { name: corpName?.name ?? `Corporation #${corporationId}`, portrait: corporationLogo(corporationId) }
+  }
+
+  // A shared ship's owner is outside the caller's registration view (RLS);
+  // their public name resolves through the world-readable directory. Both are
+  // probed together — only one of them can match.
+  const [{ data: registration }, { data: directory }] = await Promise.all([
+    supabase
+      .from('registration')
+      .select('name, character_id')
+      .eq('id', characterSelf.registration_id)
+      .maybeSingle<{ name: string; character_id: number | string | null }>(),
+    supabase
+      .from('character_directory')
+      .select('name, character_id')
+      .eq('registration_id', characterSelf.registration_id)
+      .maybeSingle<{ name: string | null; character_id: number | string | null }>(),
+  ])
+  const eveCharacterId = registration?.character_id ?? directory?.character_id ?? null
+  return {
+    name: registration?.name ?? directory?.name ?? 'Unknown character',
+    portrait: eveCharacterId == null ? null : characterPortrait(eveCharacterId),
+  }
 }
 
 // Anonymous share-link view: wheel + name + owner only. All queries run on
@@ -320,7 +222,7 @@ const SharedShipPage = async ({
         .maybeSingle<ShipRow>()
   const self = characterSelf ?? corpSelf
   // The share outlived the ship (sold, transferred, unlinked): dead link.
-  if (!self || !(await isShipType(self.type_id))) notFound()
+  if (!self || (await getSdeType(Number(self.type_id)))?.categoryID !== SHIP_CATEGORY_ID) notFound()
 
   const [{ data: characterChildren }, { data: corpChildren }] = await Promise.all([
     supabase
